Moved content service to core. Fixed crash where animations wouldnt load in game
